ThemeParkStartPoint/src/test/java/stalls/TobaccoStallTest.java
package stalls;

import org.junit.Before;
import org.junit.Test;
import people.Visitor;

import static org.junit.Assert.assertEquals;

public class TobaccoStallTest {

    TobaccoStall tobaccoStall;

    @Before
    public void setUp() throws Exception {
        tobaccoStall = new TobaccoStall("Jacks Drum", "Jack Jarvis", ParkingSpot.B1);
    }

    @Test
    public void hasName() {
        assertEquals("Jacks Drum", tobaccoStall.getName());
    }

    @Test
    public void hasRating(){ assertEquals(5, tobaccoStall.getRating()); }

    @Test
    public void hasOwner() {
        assertEquals("Jack Jarvis", tobaccoStall.getOwnerName());
    }

    @Test
    public void hasParkingSpot() {
        assertEquals(ParkingSpot.B1, tobaccoStall.getParkingSpot());
    }

    @Test
    public void hasMinAge_NotAllowed(){
        Visitor visitor = new Visitor(14, 12, 13.30);
        assertEquals(false, tobaccoStall.isAllowedTo(visitor));
    }

    @Test
    public void hasMinAge_Allowed(){
        Visitor visitor = new Visitor(19, 12, 13.30);
        assertEquals(true, tobaccoStall.isAllowedTo(visitor));
    }
}
